fix: move GuiArea to correct package 'scene.context'
